refactor: extract hook protocol into coding-agent-hooks crate
Move the agent-agnostic hook handling logic into a new workspace crate
called `coding-agent-hooks`. The name `agent-hooks` was taken on
crates.io.

The new crate provides:
- AgentKind enum with optional clap::ValueEnum (behind `clap` feature)
- HookProtocol trait abstracting agent-specific JSON formats
- Hook input/output types with an Effect enum (Allow/Ask/Deny)
- Protocol implementations for all 6 agents
- Tool name normalization and permission mode resolution tables

The clash crate now depends on coding-agent-hooks and re-exports its
types through crate::hooks and crate::agents. Clash-specific extensions
(typed_tool_input via ToolUseHookInputExt trait, is_interactive_tool,
SelectItem impl) remain in the clash crate.

Replaces claude_settings::PermissionRule in hook output types with the
new coding_agent_hooks::output::Effect enum, which serializes
identically (allow/ask/deny).
