+++ coding-agent-hooks/src/agents/gemini.rs
@@ -7,9 +7,9 @@
 use anyhow::Result;
 use serde_json::Value;
 
-use super::protocol::{HookProtocol, json_str};
 use super::{AgentKind, resolve_tool_name};
-use crate::hooks::ToolUseHookInput;
+use crate::input::ToolUseHookInput;
+use crate::protocol::{HookProtocol, json_str};
 
 pub struct GeminiProtocol;
